refactor(environments): split mod.rs into config/devx_core/catalog submodules
rust/src/environments/mod.rs was approaching the 1000-line CI limit,
blocking contributors adding to it in an open PR and another branch.
Splits it by responsibility: GddyEnvConfig + URL resolution helpers
(config.rs), DevX Core gateway URL resolution (devx_core.rs), and
API-catalog domain base-URL resolution (catalog.rs), with each
submodule's tests moving alongside it. mod.rs keeps only the
compiled-in registry, the shared Environments instance, and the
public resolve/listable/is_known API. Public API is unchanged via
re-exports.
